derive_butcher_struct: migrate all the code about fields in a separate module
This will heavily reduce code duplication while working with enums.
